requirements.txt
flask>=3.0.0
typer>=0.12.0
pandas>=2.2.0
numpy>=1.26.0
yfinance>=0.2.40
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
jinja2>=3.1.4
matplotlib>=3.9.0
pytz>=2024.1
python-dateutil>=2.9.0
pypdf>=5.0.0
feedparser>=6.0.11
pyyaml>=6.0.2
openpyxl>=3.1.2
plotly>=5.18.0
ibapi>=9.81.1
